+++ mlrose/algorithms/__init__.py
@@ -14,4 +14,4 @@ from .crossovers import UniformCrossOver, TSPCrossOver, OnePointCrossOver
 
 from .decay import ArithDecay, CustomSchedule, ExpDecay, GeomDecay
 from .mutators import ChangeOneMutator, DiscreteMutator, ShiftOneMutator, SwapMutator
-from .decorators import (short_name)
+from .decorators import (get_short_name)
