+++ mlrose/algorithms/decorators/__init__.py
@@ -4,4 +4,4 @@
 # Author: Genevieve Hayes (modified by Andrew Rollings)
 # License: BSD 3 clause
 
-from .short_name_decorator import short_name
+from .short_name_decorator import short_name, get_short_name
