+++ mlrose/algorithms/decorators/short_name_decorator.py
@@ -10,3 +10,7 @@ def short_name(expr):
         func.__short_name__ = expr
         return func
     return short_name_func_applicator
+
+
+def get_short_name(v):
+    return v if not hasattr(v, '__short_name__') else v.__short_name__
