+++ mlrose/runners/__init__.py
@@ -1,3 +1,8 @@
+""" Classes for running optimization problems."""
+
+# Author: Genevieve Hayes (modified by Andrew Rollings)
+# License: BSD 3 clause
+
 from .ga_runner import GARunner
 from .rhc_runner import RHCRunner
 from .sa_runner import SARunner
