+++ mlrose/runners/mimic_runner.py
@@ -1,6 +1,5 @@
 import mlrose
-from mlrose import short_name
-from mlrose.opt_probs import DiscreteOpt
+from mlrose.algorithms.decorators import short_name
 from mlrose.runners._runner_base import _RunnerBase
 
 """
